fix: login-page component
